fixed init

+++ src/GPM/Init.hs
@@ -11,13 +11,15 @@ module GPM.Init
   (init)
 where
 
-import           Protolude
+import           Protolude     hiding (die,fold)
 import           Turtle
 
-import           GPM.Helpers (debug_)
-import qualified GPM.Docs    as Docs
-import qualified GPM.Issue   as Issue
-import qualified GPM.Review  as Review
+import qualified Control.Foldl as Fold
+
+import qualified GPM.Docs      as Docs
+import           GPM.Helpers   (debug_)
+import qualified GPM.Issue     as Issue
+import qualified GPM.Review    as Review
 
 -- | Init a repository with a new empty branch named @gpm@
 init :: IO ()
@@ -32,9 +34,14 @@ init = do
 
 -- | Create a new empty branch, fail if the branch already exists
 mkNewEmptyBranch :: Text -> IO ()
-mkNewEmptyBranch br = do
-  putText $ "create a new branch " <> br <> " (be sure the branch " <> br <> " doesn't already exists)"
-  debug_ $ "git checkout --orphan " <> br
-  echo "cleanup the branch"
-  debug_ "git rm --cached -r ."
+mkNewEmptyBranch br = sh $ do
+  clean <- fold (inshell "git status --porcelain" empty) Fold.null
+  if clean
+    then liftIO $ do
+      putText $ "create a new branch " <> br <> " (be sure the branch " <> br <> " doesn't already exists)"
+      debug_ $ "git checkout --orphan " <> br
+      echo "cleanup the branch"
+      debug_ "git rm --cached -r ."
+      debug_ "git clean -fd"
+    else die "Please take care of pending modification and untracked files (you can use git stash --all)"
